Mdct 461 Create ComplexRate component (shared for AIFHH and IUHH) (#1385)
* refactor in progress - combine duplicated parts

* fix readonly fields

* cleanup for pr

* add tests and cleanup

* wip everything works except iiuh optionals

* fixup missing fields, fix tests
