--- services/ui-src/src/components/IUHHRate/index.test.tsx
@@ -1,9 +1,0 @@
-import { renderWithHookForm } from "utils/testUtils/reactHookFormRenderer";
-
-describe("Test the Rate component", () => {
-  beforeEach(() => {
-    renderWithHookForm(<></>);
-  });
-
-  test("Check that component renders", () => {});
-});
