+++ services/ui-src/src/measures/2022/AIFHH/data.ts
@@ -1,5 +1,42 @@
 import { DataDrivenTypes } from "measures/2022/CommonQuestions/types";
 
+const measureName = "AIFHH";
+
+const inputFieldNames = [
+  "Number of Enrollee Months",
+  "Number of Short-Term Admissions",
+  "Short-Term Admissions per 1,000 Enrollee Months",
+  "Number of Medium-Term Admissions",
+  "Medium-Term Admissions per 1,000 Enrollee Months",
+  "Number of Long-Term Admissions",
+  "Long-Term Admissions per 1,000 Enrollee Months",
+];
+
+// Rate structure by index in row
+const ndrFormulas = [
+  // Short-Term Admissions per 1,000 Enrollee Months
+  {
+    num: 1,
+    denom: 0,
+    rate: 2,
+    mult: 1000,
+  },
+  // Medium-Term Admissions per 1,000 Enrollee Months
+  {
+    num: 3,
+    denom: 0,
+    rate: 4,
+    mult: 1000,
+  },
+  // Long-Term Admissions per 1,000 Enrollee Months
+  {
+    num: 5,
+    denom: 0,
+    rate: 6,
+    mult: 1000,
+  },
+];
+
 export const categories = [];
 export const qualifiers = [
   "Ages 18 to 64",
@@ -22,6 +59,9 @@ export const data: DataDrivenTypes.PerformanceMeasure = {
   questionSubtext: [
     "The following three rates are reported across four age groups (ages 18 to 64, ages 65 to 74, ages 75 to 84, and age 85 and older):",
   ],
+  measureName,
+  inputFieldNames,
+  ndrFormulas,
   categories,
   qualifiers,
 };
