+++ services/ui-src/src/measures/2022/CommonQuestions/OptionalMeasureStrat/index.tsx
@@ -7,6 +7,7 @@ import { TopLevelOmsChildren } from "./omsNodeBuilder";
 import { useCustomRegister } from "hooks/useCustomRegister";
 import { useEffect } from "react";
 import { useFormContext } from "react-hook-form";
+import { ndrFormula } from "types";
 
 interface OmsCheckboxProps {
   /** name for react-hook-form registration */
@@ -48,6 +49,9 @@ export const buildOmsCheckboxes = ({
 };
 
 interface BaseProps extends Types.Qualifiers, Types.Categories {
+  measureName?: string;
+  inputFieldNames?: string[];
+  ndrFormulas?: ndrFormula[];
   /** string array for perfromance measure descriptions */
   performanceMeasureArray?: Types.RateFields[][];
   IUHHPerformanceMeasureArray?: Types.complexRateFields[][];
@@ -115,6 +119,9 @@ export const OptionalMeasureStrat = ({
   AIFHHPerformanceMeasureArray,
   qualifiers = [],
   categories = [],
+  measureName,
+  inputFieldNames,
+  ndrFormulas,
   data,
   calcTotal = false,
   adultMeasure,
@@ -174,6 +181,9 @@ export const OptionalMeasureStrat = ({
           rateReadOnly,
           calcTotal,
           qualifiers,
+          measureName,
+          inputFieldNames,
+          ndrFormulas,
           categories,
           rateMultiplicationValue,
           customMask,
